# Speed up `tests/samplers_tests/test_nsgaii.py::test_fast_non_dominated_sort_with_constraints`