--- a/optuna/samplers/nsgaii/_sampler.py
+++ b/optuna/samplers/nsgaii/_sampler.py
@@ -345,9 +345,12 @@ class NSGAIISampler(BaseSampler):
         population: List[FrozenTrial],
         directions: List[optuna.study.StudyDirection],
     ) -> List[List[FrozenTrial]]:
+        for _trial in population:
+            _constraints = _trial.system_attrs.get(_CONSTRAINTS_KEY)
+            if np.any(np.isnan(_constraints)):
+                raise ValueError("NaN is not acceptable as constraint value.")
         dominated_count: DefaultDict[int, int] = defaultdict(int)
         dominates_list = defaultdict(list)
-
         dominates = _dominates if self._constraints_func is None else _constrained_dominates
 
         for p, q in itertools.combinations(population, 2):